add testing module
