+++ bach/bach/_testing/__init__.py
@@ -1,0 +1,1 @@
+from bach._testing.utils import assert_equals_data
