+++ bach/bach/testing.py
@@ -1,0 +1,5 @@
+
+"""
+Bach public testing utility functions.
+"""
+from bach._testing import assert_equals_data
