+++ bach/setup.cfg
@@ -53,7 +53,7 @@ dev =
 
 [options.packages.find]
 where = .
-exclude = tests, tests.*
+exclude = tests, tests.*, bach._testing, bach._testing.*
 
 [options.package_data]
 * = py.typed
